+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Feedback: User email address as the from address (HUB-190)
 * Display search filters only when more than one type of result (HUB-202)
 * Redirect external active degree programme calls to front page (HUB-184)
+* Added paragraphs to themes (HUB-201)
 
 
 ## 1.2